requirements.txt
requests==2.24.0
discord.py==1.5.1
python-dotenv==0.14.0
requests-cache==0.5.2
websockets==8.1